remove int conversion in funtion get_severe_cases

+++ src/estimator.py
@@ -27,7 +27,7 @@ def estimator(data):
 
     def get_severe_cases(estimate_type="impact"):
         cases = convid_cases if estimate_type == "impact" else severe_cases
-        return int(cases * 0.15)
+        return cases * 0.15
 
     def get_hospital_beds(estimate_type="impact"):
         available_beds = data["totalHospitalBeds"] * 0.35
